Refactor/move LogEmitter to make it reusable by operators
