+++ processor/logstransformprocessor/processor.go
@@ -18,6 +18,7 @@ import (
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/adapter"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/pipeline"
 )
 
@@ -29,7 +30,7 @@ type logsTransformProcessor struct {
 
 	pipe          *pipeline.DirectedPipeline
 	firstOperator operator.Operator
-	emitter       *adapter.LogEmitter
+	emitter       *helper.LogEmitter
 	converter     *adapter.Converter
 	fromConverter *adapter.FromPdataConverter
 	shutdownFns   []component.ShutdownFunc
@@ -44,7 +45,7 @@ func newProcessor(config *Config, nextConsumer consumer.Logs, logger *zap.Logger
 
 	baseCfg := p.config.BaseConfig
 
-	p.emitter = adapter.NewLogEmitter(p.logger.Sugar())
+	p.emitter = helper.NewLogEmitter(p.logger.Sugar())
 	pipe, err := pipeline.Config{
 		Operators:     baseCfg.Operators,
 		DefaultOutput: p.emitter,
